requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
numpy>=1.24.0
rarfile>=4.0
